[Chore] #8 - UIBezierPath extension 이동
