+++ Motivoo-iOS/Motivoo-iOS/Feature/Home/Views/HomeCircularProgressView.swift
@@ -10,15 +10,6 @@ import UIKit
 import SnapKit
 import Then
 
-extension UIBezierPath {
-    static func point(atPercent percent: CGFloat, withRadius radius: CGFloat, andCenter center: CGPoint) -> CGPoint {
-        let angle = 2 * CGFloat.pi * percent - CGFloat.pi / 2
-        let x = center.x + radius * cos(angle)
-        let y = center.y + radius * sin(angle)
-        return CGPoint(x: x, y: y)
-    }
-}
-
 final class HomeCircularProgressView: BaseView {
     
     // MARK: - Properties
